Create m.file
This code reads the input image into the variable img, then converts the image to a black and white image. It utilizes edge detection technique to detect tree edges and displays the detected edges in the image. Note that this code is just a simple example and might require further adjustments and modifications to achieve more accurate results.